pyproject.toml
[project]
name = "daylight"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "naphal", email = "naphal@naphal.net" }
]
dependencies = [
    "pysigma>=0.11.10",
    "pysigma-backend-splunk>=1.1.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/daylight"]
